update.sh: Add NXP wifi driver to package build
Include the possibility to add the NXP wifi driver as out-of-tree source
to the package build. If added, the driver will be build during the
kernel builds with the options as configured in the driver makefile. Be
sure to checkout the correct branch for the driver before building the
kernel package.

Signed-off-by: Frank Erdrich <f.erdrich@kunbus.com>